Added onOffice estate reads
